Removed app directory to have a flatter structure
