+++ api/src/functions.py
@@ -1,0 +1,6 @@
+import os
+
+audience = os.getenv("API_URL")
+api_key = os.environ.get("API_KEY")
+
+headers = {"X-API-KEY": api_key}
